:recycle: refactor: entity를 각 도메인네 entity 폴더로 이동
