+++ backend/src/app/group-article/entity/article.entity.ts
