+++ backend/src/app/group-article/entity/group-category.entity.ts
